src/main/java/org/example/lesson4/Triangle.java
package org.example.lesson4;


import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.testng.annotations.Test;

import java.math.BigDecimal;
import java.math.RoundingMode;

@Data
@NoArgsConstructor

public class Triangle {
    private int a;
    private int b;
    private int c;

    public Triangle(int a, int b, int c) {
        this.a = a;
        this.b = b;
        this.c = c;
    }

    public static double round(double value, int places) {
        if (places < 0) throw new IllegalArgumentException();

        BigDecimal bd = BigDecimal.valueOf(value);
        bd = bd.setScale(places, RoundingMode.HALF_UP);
        return bd.doubleValue();
    }

    public boolean isValid() {
        double maxSide = Math.max (a, Math.max(b, c));
        return maxSide < (a + b + c - maxSide);
    }

    public boolean hasPositiveSides() {
        return (a > 0 && b > 0 && c > 0);
    }

    public double countSquare() {
        if (!hasPositiveSides()) {
            throw new IllegalArgumentException("The sides must be positive");
        }
        if (!isValid()) {
            throw new IllegalArgumentException("The triangle must be valid");
        }
        double p = (double) (a + b + c)/2;
        double square = round(Math.sqrt(p*(p-a)*(p-b)*(p-c)),2);
        return square;
    }


}
